+++ lib/plausible/stats/filter_suggestions.ex
@@ -180,6 +180,10 @@ defmodule Plausible.Stats.FilterSuggestions do
     |> wrap_suggestions()
   end
 
+  # Deprecated (buggy) and will be replaced with a new endpoint (see
+  # custom_prop_value_filter_suggestion/4). Will only be kept around
+  # for some time until the dashboards get updated to query the new
+  # endpoint. Tests have already been adjusted to the new endpoint.
   def filter_suggestions(site, query, "prop_value", filter_search) do
     filter_query = if filter_search == nil, do: "%", else: "%#{filter_search}%"
 
@@ -267,6 +271,40 @@ defmodule Plausible.Stats.FilterSuggestions do
     |> limit(25)
     |> ClickhouseRepo.all()
     |> Enum.filter(fn suggestion -> suggestion != "" end)
+    |> wrap_suggestions()
+  end
+
+  def custom_prop_value_filter_suggestions(site, query, prop_key, filter_search) do
+    filter_query = if filter_search == nil, do: "%", else: "%#{filter_search}%"
+
+    none_q =
+      from(
+        e in base_event_query(
+          site,
+          Query.remove_top_level_filters(query, ["event:props:#{prop_key}"])
+        ),
+        select: "(none)",
+        where: not has_key(e, :meta, ^prop_key),
+        limit: 1
+      )
+
+    search_q =
+      from(e in base_event_query(site, query),
+        select: get_by_key(e, :meta, ^prop_key),
+        where:
+          has_key(e, :meta, ^prop_key) and
+            fragment(
+              "? ilike ?",
+              get_by_key(e, :meta, ^prop_key),
+              ^filter_query
+            ),
+        group_by: get_by_key(e, :meta, ^prop_key),
+        order_by: [desc: fragment("count(*)")],
+        limit: 25
+      )
+
+    ClickhouseRepo.all(none_q)
+    |> Kernel.++(ClickhouseRepo.all(search_q))
     |> wrap_suggestions()
   end