+++ test/plausible_web/controllers/api/stats_controller/suggestions_test.exs
@@ -424,7 +424,7 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
              ]
     end
 
-    test "returns suggestions found in time frame", %{
+    test "returns prop key suggestions found in time frame", %{
       conn: conn,
       site: site
     } do
@@ -442,11 +442,39 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
       ])
 
       conn =
-        get(conn, "/api/stats/#{site.domain}/suggestions/prop_key?period=day&date=2022-01-01")
+        get(conn, "/api/stats/#{site.domain}/suggestions/prop_key?period=day&date=2022-01-01&q=")
 
       assert json_response(conn, 200) == [
                %{"label" => "author", "value" => "author"},
                %{"label" => "logged_in", "value" => "logged_in"}
+             ]
+    end
+
+    test "returns prop key suggestions by a search string", %{
+      conn: conn,
+      site: site
+    } do
+      populate_stats(site, [
+        build(:pageview,
+          "meta.key": ["author", "logged_in"],
+          "meta.value": ["Uku Taht", "false"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          "meta.key": ["dark_mode"],
+          "meta.value": ["true"],
+          timestamp: ~N[2022-01-02 00:00:00]
+        )
+      ])
+
+      conn =
+        get(
+          conn,
+          "/api/stats/#{site.domain}/suggestions/prop_key?period=day&date=2022-01-01&q=aut"
+        )
+
+      assert json_response(conn, 200) == [
+               %{"label" => "author", "value" => "author"}
              ]
     end
 
@@ -573,12 +601,49 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
       conn =
         get(
           conn,
-          "/api/stats/#{site.domain}/suggestions/prop_value?period=day&date=2022-01-01&filters=#{filters}"
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/author?period=day&date=2022-01-01&filters=#{filters}&q="
         )
 
       assert json_response(conn, 200) == [
                %{"label" => "(none)", "value" => "(none)"},
                %{"label" => "Uku Taht", "value" => "Uku Taht"},
+               %{"label" => "Marko Saric", "value" => "Marko Saric"}
+             ]
+    end
+
+    test "returns (none) value as the first suggestion even when a search string is provided",
+         %{conn: conn, site: site} do
+      populate_stats(site, [
+        build(:pageview,
+          "meta.key": ["author"],
+          "meta.value": ["Uku Taht"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          "meta.key": ["author"],
+          "meta.value": ["Marko Saric"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          "meta.key": ["author"],
+          "meta.value": ["Marko Saric"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          timestamp: ~N[2022-01-01 00:00:00]
+        )
+      ])
+
+      filters = Jason.encode!(%{props: %{author: "!(none)"}})
+
+      conn =
+        get(
+          conn,
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/author?period=day&date=2022-01-01&filters=#{filters}&q=Mar"
+        )
+
+      assert json_response(conn, 200) == [
+               %{"label" => "(none)", "value" => "(none)"},
                %{"label" => "Marko Saric", "value" => "Marko Saric"}
              ]
     end
@@ -610,12 +675,48 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
       conn =
         get(
           conn,
-          "/api/stats/#{site.domain}/suggestions/prop_value?period=day&date=2022-01-01&filters=#{filters}"
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/author?period=day&date=2022-01-01&filters=#{filters}"
         )
 
       assert json_response(conn, 200) == [
                %{"label" => "Uku Taht", "value" => "Uku Taht"},
                %{"label" => "Marko Saric", "value" => "Marko Saric"}
+             ]
+    end
+
+    test "returns prop value suggestions with multiple custom property filters in query", %{
+      conn: conn,
+      site: site
+    } do
+      populate_stats(site, [
+        build(:pageview,
+          "meta.key": ["author", "browser_language"],
+          "meta.value": ["Uku Taht", "en-US"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          "meta.key": ["author", "browser_language"],
+          "meta.value": ["Uku Taht", "en-US"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview,
+          "meta.key": ["author", "browser_language"],
+          "meta.value": ["Marko Saric", "de-DE"],
+          timestamp: ~N[2022-01-01 00:00:00]
+        ),
+        build(:pageview, timestamp: ~N[2022-01-01 00:00:00])
+      ])
+
+      filters = Jason.encode!(%{props: %{browser_language: "!(none)", author: "Uku Taht"}})
+
+      conn =
+        get(
+          conn,
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/browser_language?period=day&date=2022-01-01&filters=#{filters}"
+        )
+
+      assert json_response(conn, 200) == [
+               %{"label" => "en-US", "value" => "en-US"}
              ]
     end
 
@@ -644,7 +745,7 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
       conn =
         get(
           conn,
-          "/api/stats/#{site.domain}/suggestions/prop_value?period=all&date=CLEVER_SECURITY_RESEARCH&filters=#{filters}"
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/author?period=all&date=CLEVER_SECURITY_RESEARCH&filters=#{filters}"
         )
 
       assert json_response(conn, 400) == %{
@@ -1008,7 +1109,7 @@ defmodule PlausibleWeb.Api.StatsController.SuggestionsTest do
       value_conn =
         get(
           conn,
-          "/api/stats/#{site.domain}/suggestions/prop_value?period=day&date=2022-01-01&with_imported=true&filters=#{filters}"
+          "/api/stats/#{site.domain}/suggestions/custom-prop-values/url?period=day&date=2022-01-01&with_imported=true&filters=#{filters}"
         )
 
       assert json_response(value_conn, 200) == [